Add code for button LED and button LED toggle.

+++ doc/digital-input/button/button-internal-pull-up-resistor-toggle-count-debounce/code/button-internal-pull-up-resistor-toggle-count-debounce/button-internal-pull-up-resistor-toggle-count-debounce.ino
@@ -7,8 +7,8 @@
 // Assign button alias to pin 2
 #define button 2
 
-static unsigned long timeStamp;
-unsigned long debounceDelay = 10; // (ms)
+static uint32_t timeStamp;
+uint32_t debounceDelay = 10; // (ms)
 
 void setup() {
   // Start Serial Monitor
